+++ CHANGELOG.md
@@ -35,4 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Hotkeys
 - Sentry support
 - News menu item
+
+### Changed
+- Signature is now customizable
 - Support multiple config, one per username